Extract library
